+++ UIComponents.Abstractions/Enums/SortOrder.cs
@@ -1,0 +1,7 @@
+﻿namespace UIComponents.Abstractions.Enums;
+
+public enum SortOrder
+{
+    Asc = 0,
+    Desc = 1,
+}