+++ UIComponents.Abstractions/Interfaces/Tables/IUICTableColumn.cs
@@ -1,6 +1,14 @@
-﻿namespace UIComponents.Abstractions.Interfaces.Tables;
+﻿using UIComponents.Abstractions.Enums;
+
+namespace UIComponents.Abstractions.Interfaces.Tables;
 
 public interface IUICTableColumn : IUIComponent, IUICConditionalRender
 {
     public string Type { get; }
+}
+
+public interface IUICSortableTableColumn : IUICTableColumn
+{
+    public string ColumnName { get; }
+    public SortOrder? SortOrder { get; }
 }