+++ UIComponents.Web.Tests/UIComponents/CHANGELOG.md
@@ -1,3 +1,10 @@
+# 2024-11-13 (1.3.09)
+- Minor fix sorting jsgrid
+- loading script from javascript
+- loading style from javascript
+- Css rework
+
+
 # 2024-11-08 (1.3.08)
 - Minor fix selectlistitems from htmlstorage
 - Minor fix on disposing partial